feat: Implement user permission checks and metadata handling in delete and update logic
- Added user ID retrieval and permission validation in DeleteLogic for images, articles, and comments to ensure only authorized users can perform deletions.
- Enhanced DeleteArticleLogic and DeleteCommentLogic to utilize user metadata for permission checks, improving security.
- Updated UpdateArticleLogic to incorporate user metadata handling, ensuring proper context during article updates.
- Refactored related repository methods to accept context with user metadata, streamlining the logic and enhancing maintainability.

These changes improve the security and integrity of delete and update operations across the application by enforcing user permissions and utilizing metadata effectively.

+++ rpc/TechSphere/Comment/internal/logic/deletecommentlogic.go
@@ -2,10 +2,13 @@ package logic
 
 import (
 	"context"
+	"fmt"
+	"strconv"
 
 	"coderhub/rpc/ImageRelation/imageRelation"
 	"coderhub/rpc/TechSphere/Comment/comment"
 	"coderhub/rpc/TechSphere/Comment/internal/svc"
+	"coderhub/shared/MetaData"
 
 	"github.com/zeromicro/go-zero/core/logx"
 )
@@ -26,11 +29,24 @@ func NewDeleteCommentLogic(ctx context.Context, svcCtx *svc.ServiceContext) *Del
 
 // DeleteComment 删除评论
 func (l *DeleteCommentLogic) DeleteComment(in *comment.DeleteCommentRequest) (*comment.DeleteCommentResponse, error) {
+	// 权限校验
+	var (
+		userId string
+		err    error
+	)
+	if userId, err = MetaData.GetUserMetaData(l.ctx); err != nil {
+		return nil, err
+	}
+
+	if userId != strconv.FormatInt(in.UserId, 10) {
+		return nil, fmt.Errorf("非法操作")
+	}
+
 	if err := l.svcCtx.CommentRepository.Delete(l.ctx, in.CommentId); err != nil {
 		return nil, err
 	}
 	// 删除图片关联
-	_, err := l.svcCtx.ImageRelationService.BatchDeleteRelation(l.ctx, &imageRelation.BatchDeleteRelationRequest{
+	_, err = l.svcCtx.ImageRelationService.BatchDeleteRelation(l.ctx, &imageRelation.BatchDeleteRelationRequest{
 		Ids: []int64{in.CommentId},
 	})
 	if err != nil {
